+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "build": "tsc -p .",
     "watch": "tsc -p . --watch",
-    "test": "mocha"
+    "test": "mocha --recursive"
   },
   "license": "MIT",
   "author": "erri120 <erri120@protonmail.com>",
