+++ src/rpgmakerTypes/ProjectPathRegistry.ts
@@ -1,4 +1,4 @@
-import { Path } from "../io/Path";
+import Path from "../io/Path";
 
 export interface ProjectPathRegistry {
   top: Path,
@@ -37,16 +37,6 @@ export interface ProjectPathRegistry {
   js: Path,
   movies: Path,
   save: Path
-}
-
-export interface TemplatePathRegistry {
-  top: Path,
-
-  // chromium portable native client, Windows only
-  pnacl: Path,
-
-  // nwjs.app, OSX only
-  nwjs_app: Path
 }
 
 export function createProjectPathRegistry(p: Path): ProjectPathRegistry {
@@ -90,18 +80,6 @@ export function createProjectPathRegistry(p: Path): ProjectPathRegistry {
     js: p.join("js"),
     movies: p.join("movies"),
     save: p.join("save")
-  };
-
-  return res;
-}
-
-export function createTemplatePathRegistry(p: Path): TemplatePathRegistry {
-  const res: TemplatePathRegistry = {
-    top: p,
-
-    pnacl: p.join("pnacl"),
-
-    nwjs_app: p.join("nwjs.app"),
   };
 
   return res;
