+++ src/rpgmakerTypes/RPGMakerPlatform.ts
@@ -1,0 +1,9 @@
+enum RPGMakerPlatform {
+  Windows = "Windows",
+  OSX = "OSX",
+  Linux = "Linux",
+  Browser = "Browser",
+  Mobile = "Mobile"
+}
+
+export default RPGMakerPlatform;
